+++ eleeye/book.h
@@ -30,19 +30,19 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
 struct BookStruct {
   union {
     uint32_t dwZobristLock;
-    int nPtr;
+    int64_t nPtr;
   };
   uint16_t wmv, wvl;
 }; // bk
 
-inline int BOOK_POS_CMP(const BookStruct &bk, const PositionStruct &pos) {
+inline int64_t BOOK_POS_CMP(const BookStruct &bk, const PositionStruct &pos) {
   return bk.dwZobristLock < pos.zobr.dwLock1 ? -1 :
       bk.dwZobristLock > pos.zobr.dwLock1 ? 1 : 0;
 }
 
 struct BookFileStruct {
   FILE *fp;
-  int nLen;
+  int64_t nLen;
   bool Open(const char *szFileName, bool bEdit = false) {
     fp = fopen(szFileName, bEdit ? "r+b" : "rb");
     if (fp == NULL) {
@@ -56,17 +56,17 @@ struct BookFileStruct {
   void Close(void) const {
     fclose(fp);
   }
-  void Read(BookStruct &bk, int nPtr) const {
+  void Read(BookStruct &bk, int64_t nPtr) const {
     fseek(fp, nPtr * sizeof(BookStruct), SEEK_SET);
     fread(&bk, sizeof(BookStruct), 1, fp);
   }
-  void Write(const BookStruct &bk, int nPtr) const {
+  void Write(const BookStruct &bk, int64_t nPtr) const {
     fseek(fp, nPtr * sizeof(BookStruct), SEEK_SET);
     fwrite(&bk, sizeof(BookStruct), 1, fp);
   }
 };
 
 // 获取开局库着法
-int GetBookMoves(const PositionStruct &pos, const char *szBookFile, BookStruct *lpbks);
+int64_t GetBookMoves(const PositionStruct &pos, const char *szBookFile, BookStruct *lpbks);
 
 #endif
